+++ src/main/kotlin/io/snyk/plugin/ui/UIUtils.kt
@@ -30,8 +30,11 @@ import javax.swing.JEditorPane
 import javax.swing.JLabel
 import javax.swing.JPanel
 import javax.swing.JScrollPane
+import javax.swing.JTree
 import javax.swing.ScrollPaneConstants
 import javax.swing.text.html.HTMLDocument
+import javax.swing.tree.DefaultMutableTreeNode
+import javax.swing.tree.TreePath
 
 fun boldLabel(title: String): JLabel {
     val label = JLabel(title)
@@ -361,3 +364,10 @@ fun txtToHtml(s: String): String {
 }
 
 fun getDisabledIcon(originalIcon: Icon?): Icon? = originalIcon?.let { IconLoader.getDisabledIcon(it) }
+
+fun expandTreeNodeRecursively(tree: JTree, node: DefaultMutableTreeNode) {
+    tree.expandPath(TreePath(node.path))
+    node.children().asSequence().forEach {
+        expandTreeNodeRecursively(tree, it as DefaultMutableTreeNode)
+    }
+}